backend/requirements.txt
Django>=4.0,<5.0
django_cors_headers>=3.11
django-extensions>=3.1.5
django-filter>=2.4.0
djangorestframework>=3.13
psycopg2-binary>=2.9
